Restructure headers module
