+++ fitsio/src/headers/constants.rs
@@ -1,0 +1,4 @@
+// FLEN_VALUE
+pub(super) const MAX_VALUE_LENGTH: usize = 71;
+// FLEN_COMMENT
+pub(super) const MAX_COMMENT_LENGTH: usize = 73;
